// Threshold-BackEnd-main/src/enums/users.enum.ts
export enum UserRole {
    ADMIN = "admin",
    ACADEMY_ADMIN = "academy-admin",
    ACADEMY_COORDINATOR = "academy-coordinator",
    MEMBER = "member",
    COACH = "coach",
    ADMIN_MANAGER = "admin-manager",
    CLUB_ADMIN = "club-admin",
    
    CLUB_EXECUTIVE_MANAGER = "club-executive-manager",
    CLUB_TECHNICAL_DIRECTOR = "club-technical-director",
    CLUB_SPORT_PROFILE_MANAGER = "club-sport-profile-manager",

    CLUB_COACH_HEAD = "club-coach-head",
    CLUB_COACH_ASSISTANT = "club-coach-assistant",
    CLUB_COACH_TRAINER = "club-coach-trainer",
    CLUB_COACH_ANALYST = "club-coach-analyst",

    CLUB_ADMIN_ASSISTANT = "club-admin-assistant",
    CLUB_ADMIN_MANAGER = "club-admin-manager",

}

export enum Language {
    ENGLISH = "english",
    ARABIC = "arabic",
}